feat: add image deletion functionality and update game management - Implemented delete method in ImageService to handle image deletions. - Updated GamesController to include image deletion when a game is deleted. - Enhanced updateGame method to delete old images when a new image is uploaded. - Added DeleteGameDto to manage deletion requests. - Modified game schema to include oldImage field for tracking previous images. - Updated frontend to handle image uploads and deletions in the game creation and editing forms. - Introduced CustomAlertDialog for confirmation before deletion actions. - Refactored image modal components for better image handling and cropping. - Improved API fetch utility to handle authorization tokens and form data. - Added new alert dialog and button components for consistent UI.

+++ api_nest/src/modules/games/game.controller.ts
@@ -21,7 +21,12 @@ import { Request, Response } from 'express';
 
 import { GamesService } from './game.service';
 import { ZodValidationPipe } from '../../models/zod.pipe';
-import { CreateGame, CreateGameDto, UpdateGameDto } from './game.schema';
+import {
+  CreateGame,
+  CreateGameDto,
+  DeleteGameDto,
+  UpdateGameDto,
+} from './game.schema';
 import { convertBytesToUuid } from '../../common/utils/uuid.util';
 import { AuthGuard } from '../auth/auth.guard';
 import { ImageService } from '../../integration/imageModule/image.service';
@@ -139,16 +144,46 @@ export class GamesController {
 
   @UseGuards(AuthGuard)
   @Patch(':id')
+  @UseInterceptors(FileInterceptor('file'))
   async updateGame(
     @Param('id') id: string,
     @Body() body: UpdateGameDto,
     @Res() response: Response,
+    @Req() request: Request,
+    @UploadedFile()
+    file: Express.Multer.File,
   ) {
+    const authorization = request.headers['authorization'];
+    if (!authorization) {
+      this.logger.error(
+        `[${this.createGame.name}] ${HttpStatus.UNAUTHORIZED} - Authorization header missing`,
+      );
+      return response.status(HttpStatus.UNAUTHORIZED).json({
+        message: 'Authorization header missing',
+      });
+    }
+
+    // check if the image has changed
+    let image = '';
+    if (file) {
+      image = await this.imageService.create({
+        authorization,
+        file,
+      });
+      await this.imageService.delete({
+        authorization,
+        image_url: body.oldImage!,
+      });
+    } else if (body.image) {
+      image = body.image;
+    }
+
     await this.gamesService.patch({
       id,
       name: body.name,
       description: body.description,
       price: body.price,
+      image: image,
     });
 
     this.logger.log(
@@ -162,8 +197,24 @@ export class GamesController {
 
   @UseGuards(AuthGuard)
   @Delete(':id')
-  async deleteGame(@Param('id') id: string, @Res() response: Response) {
+  async deleteGame(
+    @Param('id') id: string,
+    @Body() body: DeleteGameDto,
+    @Res() response: Response,
+    @Req() request: Request,
+  ) {
+    const authorization = request.headers['authorization'];
+    if (!authorization) {
+      this.logger.error(
+        `[${this.createGame.name}] ${HttpStatus.UNAUTHORIZED} - Authorization header missing`,
+      );
+      return response.status(HttpStatus.UNAUTHORIZED).json({
+        message: 'Authorization header missing',
+      });
+    }
+
     await this.gamesService.delete({ id });
+    await this.imageService.delete({ authorization, image_url: body.image });
 
     this.logger.log(
       `[${this.deleteGame.name}] ${HttpStatus.OK} - Game with id: ${id} deleted successfully`,
